Two small changes to server.py that allow it to run on python2.2

+++ server.py
@@ -1,3 +1,4 @@
+from __future__ import generators # add yield for python2.2
 import re, struct, socket, select, traceback, time
 if not globals().get('skip_imports'):
     import ssnet, helpers, hostwatch
@@ -43,7 +44,7 @@ def _maskbits(netmask):
     
     
 def _shl(n, bits):
-    return n * int(2**bits)
+    return n * long(2**bits)
 
 
 def _list_routes():
